pyproject.toml
[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[project]
name = "bloqade-lanes"
version = "0.5.0-dev"
description = "Components for defining and manipulating lane-based structures in Bloqade for neutral atom quantum computing."
readme = "README.md"
authors = [
    { name = "Phillip Weinberg", email = "pweinberg@quera.com" }
]
dependencies = [
    "bloqade-circuit[cirq, qasm2, tsim]~=0.13.0",
    "bloqade-geometry~=0.5.0",
    "deprecated>=1.3.1",
    "kirin-toolchain~=0.22.6",
    "numpy>=2.2.6",
    "rustworkx>=0.17.1",
    "scipy>=1.15.3",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
cudaq = [
    "cudaq>=0.13.0",
    "qbraid-qir[squin]>=0.5.1",
]
visualization = [
    "matplotlib>=3.9.4",
    "scipy>=1.15.3",
    "pyqt5>=5.15.11;  sys_platform == 'darwin'",
    "pyqt5>=5.15.11;  sys_platform == 'linux'",
]

[tool.maturin]
manifest-path = "crates/bloqade-lanes-bytecode-python/Cargo.toml"
module-name = "bloqade.lanes.bytecode._native"
python-source = "python"
python-packages = ["bloqade"]

[tool.uv]
[dependency-groups]
doc = [
    "griffe-kirin>=0.1.0",
    "griffe-inherited-docstrings>=1.1.1",
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-material>=9.5.44",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocstrings[python]>=0.27.0",
]
dev = [
    "black>=26.0.0",
    "coverage>=7.6.4",
    "ipython>=8.29.0",
    "isort>=5.13.2",
    "jupytext>=1.18.1",
    "mypy>=1.13.0",
    "numpy>=1.26.4",
    "pre-commit>=4.0.1",
    "pyinstrument>=5.1.2",
    "pyright[nodejs]>=1.1.408",
    "pyspy>=0.1.1",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
    "rust-just>=1.36.0",
]

[tool.isort]
profile = "black"
combine_as_imports = true
multi_line_output = 3
src_paths = ["python/bloqade"]

[tool.black]
line-length = 88

[tool.ruff]
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.coverage.run]
include = ["python/bloqade/*"]

[tool.coverage.report]
exclude_also = [
    '@overload',
    '@_wraps'
]

[tool.pytest.ini_options]
testpaths = "python/tests/"
